lab1 -- array processing fixes

+++ 1/arrayprocessing/arrayprocessing.h
@@ -9,31 +9,31 @@
 typedef double (*func_ptr)(double);
 
 // Создание массива
-static inline double* arrayCreate(uint32_t size) {
+static inline double* arrayCreate(size_t size) {
 	double* A = (double*)malloc(sizeof(double) * size);
 	return A;
 }
 
 // Инициализация массива случайными числами
-static inline void arrayRandInit(double* A, uint32_t size) {
-	for (uint32_t i = 0; i != size; ++i)
+static inline void arrayRandInit(double* A, size_t size) {
+	for (size_t i = 0; i != size; ++i)
 		A[i] = rand() % size;
 }
 
 // Копирование массива
-static inline void arrayCopy(double* dest, double* src, uint32_t size) {
-	for (uint32_t i = 0; i != size; ++i)
+static inline void arrayCopy(double* dest, double* src, size_t size) {
+	for (size_t i = 0; i != size; ++i)
 		dest[i] = src[i];
 }
 
 // Вывод массива на экран
-static inline void arryPrint(double* A, uint32_t size) {
-	for (uint32_t i = 0; i != size; ++i)
+static inline void arryPrint(double* A, size_t size) {
+	for (size_t i = 0; i != size; ++i)
 		printf("%g ", A[i]);
 	printf("\n");
 }
 
 // Обработка элементов массива
-void arrayProcessing(double* A, uint32_t size, func_ptr func, uint8_t threads_count);
+void arrayProcessing(double* A, size_t size, func_ptr func, uint8_t threads_count);
 
 #endif
